+++ runtime/src/lib.rs
@@ -1,6 +1,7 @@
 mod accounts;
+mod accounts_db;
 mod accounts_index;
-pub mod append_vec;
+mod append_vec;
 pub mod bank;
 pub mod bank_client;
 mod blockhash_queue;
